refactor: complete KgsCalendarApp root cleanup
